Tratamiento mejor de los errores que da el servidor de Google. Actualiza siempre el user_login. Reducido el tiempo de cache
git-svn-id: file:///home/gallir/meneame/svn/src@1981 cbe1a1ec-7179-4919-8e32-4551e8279a00

+++ branches/version3/jabber/MnmJabber.pm
@@ -188,7 +188,11 @@ sub InMessage {
 	if ( $type ne 'chat' ) {
 		print "Error type '$type' from $from\n";
 		if ( $type eq 'error' ) {
-			$Users->delete($user);
+			print "Error " . $message->GetErrorCode() . ": ". $message->GetError() . "\n";
+			if ($message->GetErrorCode() != 503) { # A common error in Gtalk
+				print "Removing $user\n";
+				$Users->delete($user);
+			}
 		}
 		return;
 	}
